Picasa import: Fix integer related security holes in PmpColumnParser
BUG=151701

Review URL: https://chromiumcodereview.appspot.com/14698014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199854 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5cfd04a2bf992779361138d4014cc67f259eb8f4

+++ chrome/browser/media_galleries/fileapi/picasa/pmp_column_reader.h
@@ -41,11 +41,11 @@ class PmpColumnReader {
  private:
   bool ParseData(const PmpFieldType expected_type, uint32* rows_read);
   // Returns the number of bytes parsed in the body, or, -1 on failure.
-  long IndexStrings();
+  int64 IndexStrings();
 
   // Source data
   scoped_ptr<uint8[]> data_;
-  size_t length_;
+  int64 length_;
 
   // Header data
   PmpFieldType field_type_;
